Update April 3rd
Changed names of tabs on Ranking page.  On add student page you can now add a new grade for that student.  We just need to implement the POST to the database.  Added placeholder for fee calculations for students on the school tab.  Fixed bug where save button disappeared on other tabs when adding new grade.