Update April 9th
POST is now working for student grade.  We have started implementing POST for students but are experiencing a bug where it wont POST(status 400).  It is saying StudentCode and StudentName cannot be null and they both have values.  StudentCode is being handled on the backend, so we are sending in a dummy studentcode in the payload.  SQL updated to add data for Student_Grades for Program_id.